Fixed some importing issues since you can use a Shader::Gradient outside a Canvas widget, where it was previously only accessible.
